backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pymongo==4.6.0
python-dotenv==1.0.0
requests>=2.32.2
psycopg2-binary>=2.9.9
scikit-learn==1.6.1
nltk==3.9.1
numpy==2.1.3
beautifulsoup4>=4.12.0
groq>=0.11.0
google-auth>=2.27.0
Pillow>=10.0.0
pytesseract>=0.3.10
SpeechRecognition>=3.10.0
twilio>=9.0.0
